move and add index.ts for withConfirmationPopup
